requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
flask==3.0.3
aiohttp==3.9.5
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
